requirements.txt
# VERILENS Dependencies

# Core frameworks
pydantic>=2.0.0
pydantic-ai>=0.0.1

# Environment configuration
python-dotenv>=1.0.0

# LLM API client
openai>=1.0.0

# Vector operations
numpy>=1.24.0
scikit-learn>=1.3.0

# PDF processing
PyPDF2>=3.0.0